entity_creation

+++ src/main/java/com/example/springesprit/repository/ChaineRestaurationRepository.java
@@ -3,5 +3,5 @@ package com.example.springesprit.repository;
 import com.example.springesprit.entity.ChaineRestauration;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChaineRestaurationRepository extends JpaRepository<ChaineRestauration, Integer> {
+public interface ChaineRestaurationRepository extends JpaRepository<ChaineRestauration, Long> {
 }
