+++ src/main/java/com/example/springesprit/repository/ChefCuisinierRepository.java
@@ -3,5 +3,5 @@ package com.example.springesprit.repository;
 import com.example.springesprit.entity.ChefCuisinier;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ChefCuisinierRepository extends CrudRepository<ChefCuisinier,Integer> {
+public interface ChefCuisinierRepository extends CrudRepository<ChefCuisinier,Long> {
 }
