+++ src/main/java/com/example/springesprit/repository/CommandeRepository.java
@@ -3,5 +3,5 @@ package com.example.springesprit.repository;
 import com.example.springesprit.entity.Commande;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommandeRepository extends JpaRepository<Commande, Integer> {
+public interface CommandeRepository extends JpaRepository<Commande, Long> {
 }
