+++ src/main/java/com/example/springesprit/repository/ComposantRepository.java
@@ -3,5 +3,5 @@ package com.example.springesprit.repository;
 import com.example.springesprit.entity.Composant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComposantRepository extends JpaRepository<Composant, Integer> {
+public interface ComposantRepository extends JpaRepository<Composant, Long> {
 }
