+++ src/main/java/com/example/springesprit/repository/DetailComposantRepository.java
@@ -3,5 +3,5 @@ package com.example.springesprit.repository;
 import com.example.springesprit.entity.DetailComposant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetailComposantRepository extends JpaRepository<DetailComposant, Integer> {
+public interface DetailComposantRepository extends JpaRepository<DetailComposant, Long> {
 }
