+++ src/main/java/com/example/springesprit/repository/RestaurantRepository.java
@@ -3,5 +3,5 @@ package com.example.springesprit.repository;
 import com.example.springesprit.entity.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantRepository extends JpaRepository<Restaurant, Integer> {
+public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
 }
